improved componentization + refactoring
