+++ backend/verbalization_interface.py
@@ -5,13 +5,13 @@ import owl_verbalizer
 logger = logging.getLogger(__name__)
 
 
-def run_verbalizer(message, repo_name, repo_type, method):
+def run_verbalizer(report, repo_name, repo_type, method):
 
     if method == "owl":
-        verbalized_explanation = owl_verbalizer.verbalize(message, repo_name, repo_type)
+        verbalized_explanation = owl_verbalizer.verbalize(report, repo_name, repo_type)
 
     elif method == "shacl":
-        verbalized_explanation = shacl_verbalizer.verbalize(message, repo_name, repo_type)
+        verbalized_explanation = shacl_verbalizer.verbalize(report, repo_name, repo_type)
 
     else:
         raise NotImplementedError()
